pubsub-receiver/requirements.txt
google-cloud-pubsub==2.13.0
google-cloud-language==2.5.0